Fix MuiThemeProvider Issue
